propsal: first version of the std::process proposal
This proposal has a lot of formating issues. This is mainly
due to the fact that the document type is not adapted to
create a C++ proposal.

The first version outlines the std::process, std::process::id
and std::this_process components. It lacks some specification
about error conditions, and is a bit scarce on its motive.

This will come in future versions of the proposal.