proposal: rename the lyx file
The lyx file contains notes about the proposal, but is not the
proposal itself. The proposal will come after we get something
good enough and stable enough to be proposed.

We may have an important editorial work to do on the proposal
in order to get a good enough document.