named_mutex: first implementation
This implementation is likely to be bugged. And in fact, it is:
the destructors of the created std::named_mutex objects are NOT
called, meaning that (under Linux) the non-anonymous shared
memory zone is still mapped. Further call to the ./ptest program
will open the shm instead of creating it.

And this is quite bad.

I don't know why the bug occurs. It might be a strange artefact
of the fork or of the exit() call although I'm not sure about
that.