process: introduce terminate_flag
The terminate_flag flag is used to tell the process class
what should the termination behavior be. It allows the user
to select whether he wants to terminate the child, the parent,
both or none after the fork() call took place (and, in case
of terminate_child, after the child callable has been
executed).

This is necessary to work around an difficult issue with
std::exit(), since objects with automatic storage that are
not global are not destroyed when std::exit() is called.